Refactor: Profile including signIn and SignUP
